add log and flink IT cases
