+++ paimon-core/src/main/java/org/apache/paimon/iceberg/IcebergCommitCallback.java
@@ -380,8 +380,17 @@ public class IcebergCommitCallback implements CommitCallback, TagCallback {
         expireAllBefore(snapshotId);
 
         if (metadataCommitter != null) {
-            metadataCommitter.commitMetadata(metadataPath, null);
-            metadataCommitter.commitMetadata(metadata, null);
+            switch (metadataCommitter.identifier()) {
+                case "hive":
+                    metadataCommitter.commitMetadata(metadataPath, null);
+                    break;
+                case "rest":
+                    metadataCommitter.commitMetadata(metadata, null);
+                    break;
+                default:
+                    throw new UnsupportedOperationException(
+                            "Unsupported metadata committer: " + metadataCommitter.identifier());
+            }
         }
     }
 
@@ -621,8 +630,17 @@ public class IcebergCommitCallback implements CommitCallback, TagCallback {
         }
 
         if (metadataCommitter != null) {
-            metadataCommitter.commitMetadata(metadataPath, baseMetadataPath);
-            metadataCommitter.commitMetadata(metadata, baseMetadata);
+            switch (metadataCommitter.identifier()) {
+                case "hive":
+                    metadataCommitter.commitMetadata(metadataPath, baseMetadataPath);
+                    break;
+                case "rest":
+                    metadataCommitter.commitMetadata(metadata, baseMetadata);
+                    break;
+                default:
+                    throw new UnsupportedOperationException(
+                            "Unsupported metadata committer: " + metadataCommitter.identifier());
+            }
         }
     }
